app/scraper/tesco.py
from .base_scraper import BaseScraper
from bs4 import BeautifulSoup
import logging

logger = logging.getLogger(__name__)

class TescoScraper(BaseScraper):
    def __init__(self):
        super().__init__('Tesco', delay=2.0)
        self.base_url = 'https://www.tesco.ie'
    
    def search_products(self, query: str):
        """Search Tesco for products - using mock data for legal compliance"""
        logger.info(f"Searching Tesco for: {query}")
        
        # In production, implement actual scraping here
        # For now, return realistic mock data
        return self._get_mock_products(query)
    
    def _get_mock_products(self, query: str):
        """Generate realistic mock data for Tesco"""
        import random
        
        # Comprehensive product database
        mock_data = {
            'milk': [
                {'name': 'Tesco Fresh Milk 1L', 'price': 1.25, 'unit': '1L'},
                {'name': 'Tesco Fresh Milk 2L', 'price': 2.15, 'unit': '2L'},
                {'name': 'Tesco Organic Milk 1L', 'price': 1.85, 'unit': '1L'},
                {'name': 'Tesco Skimmed Milk 1L', 'price': 1.25, 'unit': '1L'},
                {'name': 'Tesco Whole Milk 500ml', 'price': 0.75, 'unit': '500ml'},
                {'name': 'Tesco Lactose Free Milk 1L', 'price': 1.95, 'unit': '1L'}
            ],
            'bread': [{'name': 'Tesco White Sliced Pan', 'price': 1.10, 'unit': '800g'}, {'name': 'Tesco Wholemeal Bread', 'price': 1.45, 'unit': '800g'}],
            'eggs': [{'name': 'Tesco Free Range Eggs 12pk', 'price': 3.25, 'unit': '12 pack'}, {'name': 'Tesco Large Eggs 6pk', 'price': 1.85, 'unit': '6 pack'}],
            'butter': [{'name': 'Tesco Irish Butter 500g', 'price': 2.85, 'unit': '500g'}],
            'chicken': [{'name': 'Tesco Chicken Breast 500g', 'price': 4.25, 'unit': '500g'}],
            'bananas': [{'name': 'Tesco Bananas 1kg', 'price': 1.49, 'unit': '1kg'}],
            'apples': [{'name': 'Tesco Apples 1kg', 'price': 2.29, 'unit': '1kg'}],
            'cheese': [{'name': 'Tesco Cheddar Cheese 200g', 'price': 2.50, 'unit': '200g'}],
            'yogurt': [{'name': 'Tesco Natural Yogurt 500g', 'price': 1.75, 'unit': '500g'}],
            'pasta': [{'name': 'Tesco Spaghetti 500g', 'price': 0.85, 'unit': '500g'}],
            'rice': [{'name': 'Tesco Basmati Rice 1kg', 'price': 2.15, 'unit': '1kg'}],
            'tomatoes': [{'name': 'Tesco Cherry Tomatoes 250g', 'price': 1.95, 'unit': '250g'}],
            'onions': [{'name': 'Tesco White Onions 1kg', 'price': 1.25, 'unit': '1kg'}],
            'potatoes': [{'name': 'Tesco Rooster Potatoes 2kg', 'price': 2.49, 'unit': '2kg'}],
            'carrots': [{'name': 'Tesco Carrots 1kg', 'price': 0.89, 'unit': '1kg'}],
            'beef': [{'name': 'Tesco Beef Mince 500g', 'price': 4.50, 'unit': '500g'}],
            'pork': [{'name': 'Tesco Pork Chops 500g', 'price': 3.75, 'unit': '500g'}],
            'fish': [{'name': 'Tesco Salmon Fillet 200g', 'price': 3.99, 'unit': '200g'}],
            'cereal': [{'name': 'Tesco Cornflakes 500g', 'price': 2.25, 'unit': '500g'}],
            'coffee': [{'name': 'Tesco Instant Coffee 200g', 'price': 4.50, 'unit': '200g'}],
            'tea': [{'name': 'Tesco Tea Bags 80pk', 'price': 2.75, 'unit': '80 pack'}],
            'sugar': [{'name': 'Tesco Granulated Sugar 1kg', 'price': 1.15, 'unit': '1kg'}],
            'flour': [{'name': 'Tesco Plain Flour 1.5kg', 'price': 1.35, 'unit': '1.5kg'}],
            'oil': [{'name': 'Tesco Sunflower Oil 1L', 'price': 1.89, 'unit': '1L'}],
            'orange': [{'name': 'Tesco Oranges 1kg', 'price': 1.99, 'unit': '1kg'}],
            'lemon': [{'name': 'Tesco Lemons 500g', 'price': 1.25, 'unit': '500g'}],
            'cucumber': [{'name': 'Tesco Cucumber Each', 'price': 0.75, 'unit': 'each'}],
            'lettuce': [{'name': 'Tesco Iceberg Lettuce', 'price': 0.89, 'unit': 'each'}]
        }
        
        # Get products or create generic ones
        products = mock_data.get(query.lower())
        if not products:
            # Generate 1-2 generic products for any search term
            base_price = random.uniform(1.0, 5.0)
            products = [
                {'name': f'Tesco {query.title()}', 'price': round(base_price, 2), 'unit': 'each'},
                {'name': f'Tesco Premium {query.title()}', 'price': round(base_price * 1.3, 2), 'unit': 'each'}
            ][:random.randint(1, 2)]
        
        return [self._standardize_product({
            **product,
            'url': f'{self.base_url}/groceries/',
            'image': f'{self.base_url}/assets/images/products/placeholder.jpg'
        }) for product in products]