requirements.txt
Flask>=2.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
APScheduler>=3.10.4
gunicorn>=21.2.0
python-dotenv>=1.0.0
Flask-Limiter>=3.5.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0